Second player is now visible in the game
In the game loop (setInterval in startGame) every players now emits
the gameId and his snake as an array to the server with the message
'move'. When the servers gets this message it emits a new message
'render' to everybody but the sender with broadcast.emit. On the client
side on recieved message 'render' we parse the date and get the snake
body array of the second and the proceed to draw it with a different color.